делит код на модули
